Update package.json
Added a "description" field to briefly describe what the package does.
Removed the "files" field because it is not necessary. NPM will automatically include all files in the build/ directory when the package is published.
Added a "repository" field to specify the Git repository URL.
Added a "bugs" field to specify the URL for issue tracking.
Added a "homepage" field to specify the URL for the package's homepage or documentation.
Added some relevant "keywords".
Changed the "scripts" field to use "test" instead of "tests", and added a "publishConfig" field to set the access level of the package to public.