Update rocksdb crate to v0.19.0 to fix compilation issue
The older version does not compile properly on fedora 38 due to updates
to glibc. The version of rocksdb packaged with f38 is v7.8.3. While I
could not build rocksdb from scratch, I was able to use:
ROCKSDB_LIB_DIR="/usr/lib64"
and it picked up the `so` and I could successfulyl build.