chore: updated era-zkevm_opcode_defs dependency (#287)
# What ❔

Updated the era-zkevm_opcode_defs dependency and minimal set of other
dependencies which required update to a compatible version. This
unlocked updated sha2 and sha3 dependencies to 0.10.8.
Prover integration test run for extra safety:
https://github.com/matter-labs/zksync-era-workflows/actions/runs/6612607366

## Why ❔

To include https://github.com/matter-labs/era-zkevm_opcode_defs/pull/16
which in turn allows us to depend on up-to-date versions of sha2 and
sha3.