feat(core): adding pubdata to statekeeper and merkle tree (#259)
# What ❔

Setting the ground for bigger boojum changes. Adding boojum related
fields to the data model and types. Extending state tree to have a
concept of `StateDiffs`.

## Why ❔

We want to reduce the size of the overall boojum merge, this sets the
basis for more extensive changes. These changes are all meant to be
non-intrusive.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.

---------

Co-authored-by: “perekopskiy” <perekopskiy@dlit.dp.ua>