fix: bump zksolc from yanked version to 1.3.16 (#348)
# What ❔

Bumps the zksolc version to 1.3.16 because 1.3.15 was being used.
Assuming this is a more involved process than just bumping this, but
figured I'd PR in anyway in case it helped speed anything up. Also,
couldn't quickly figure out how to test
`docker/contract-verifier/Dockerfile` as it's broken for me for
unrelated parts.

Also fixes a typo in the init docs, I can separate it if wanted.

## Why ❔

Broken on a fresh install because of yanked version
https://github.com/matter-labs/zksolc-bin/commit/2a29ad1d8920550d9aeef195f22c8f0ca14e5626

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.