fix(house_keeper): Emit the correct circuit_id for aggregation round 2
This is us hardcoding the circuit_id to 2 for aggregation_round 2.
This fix is useless as soon as we change the config (not expected
anytime soon).

A real fix will follow where we address what is saved in database layer
and this patch will be removed altogether.

This is necessary to enable autoscaler to work (emits the correct data
for prover groups, which in turn can be picked by autoscaler).