chore: remove tslint related comments (#636)
## What ❔

This pull request aims to remove TSLint-related comments and
configurations from the codebase.

## Why ❔

The removal is necessary because the repository has transitioned to
ESLint, TSLint-related comments and configurations obsolete. This
cleanup ensures a more streamlined and consistent codebase, aligning
with the migration to ESLint for improved code quality and
maintainability.

## Checklist

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [X] Code has been formatted via `zk fmt` and `zk lint`.
- [X] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.

---------

Co-authored-by: AnastasiiaVashchuk <72273339+AnastasiiaVashchuk@users.noreply.github.com>