feat: Implemented 1 validator consensus for the main node (#554)
## What ❔

Implemented 1 validator consensus for the main node.
It is NOT enabled yet, corresponding configuration still needs to be
provided.
It is responsible for populating consensus column in the Miniblocks
table.
Consensus will be running asynchronously for now, not affecting the
miniblock production at all.

Fixes BFT-388.

## Why ❔

Reaching a consensus with just 1 validator doesn't provide much value,
however we want to start running the consensus code in production (in a
"shadow" mode) to make sure it works as expected before we start relying
on it.