feat(vm)!: Release v19 - remove allowlist (#747)
## What ❔

This upgrade contains mainly contract changes (more description of those
[here](https://github.com/matter-labs/era-contracts/pull/133)), while
the changes that impacted `zksync-era` repo mainly refer to the changes
in how the testing / contract preprocessing is done: now, the system
contracts are preprocessed into the `contracts-preprocessed` folder and
then all the artifacts should be taken from there

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.