feat(snapshot creator): Make snapshot creator fault-tolerant (#691)
## What ❔

Makes the snapshot creator tolerant to failures. The snapshot creator
now checks whether there's a partially created snapshot on start. If
there is, the creator continues producing this snapshot instead of
starting from scratch.

## Why ❔

Since the snapshot creator can run for an extended amount of time, not
having fault tolerance is unmaintainable in the long run.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `cargo spellcheck
--cfg=./spellcheck/era.cfg --code 1`.