feat: ZK Stack framework MVP (#880)
⚠️ Please read the description before reviewing. ⚠️ 

This is the ZK Stack framework MVP that represents the core
functionality of the new framework.
A lot of things are not polished yet: it'll need better
logging/reporting, interfaces can be better, naming is off at some
places (and I haven't decided on the usage of ZK Stack / zkSync there),
we need more tests and examples, no README.md, etc.

Additionally, only a few resources and tasks are implemented right now.
The reasons for that:
- Adapting the existing code to get a single interface takes time, as
most of the tasks/resources were created in an ad-hoc manner.
- I wanted to focus on the framework functionality here, the
implementations may come later as separate PRs.
- I didn't want the diff to be big.

In the current form, the framework contains a showcase of how
interacting with it looks like, and generally, it seems to work fine.
For more details on motivation and goals, please read the spec.

With that in mind, *any* feedback is appreciated, but I don't guarantee
that I'll proceed it in this PR. Here I'll focus on making sure that the
framework works and there is no major disagreement w.r.t. its design,
and then there will be N follow-up PRs bringing the missing
functionality and polishing the code.