fix(test): Provide higher min allowance in estimation (#923)
## What ❔

During the estimation of transaction, the `minAllowance` is set by the
user to the paymaster (and then it is fully transferred to the
paymaster). Note, however, that the amount of funds to be paid for this
operation depends on the size of `minAllowance` (and may also endure
additional costs if the write to the slot is an initial one).

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.