docs: Fix Typo  (#908)
## What ❔

Simple one letter typo fixes to help me understand how GitHub works and
to contribute to ZkSync.
## Why ❔

To improve readability for users and beginners. Also a way for to me to
begin contributing to projects i care about and learn how to use GitHub.
<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ Y] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ Y] Tests for the changes have been added / updated.
- [ Y] Documentation comments have been added / updated.
- [ Y] Code has been formatted via `zk fmt` and `zk lint`.
- [ Y] Spellcheck has been run via `zk spellcheck`.

---------

Co-authored-by: Fedor Sakharov <fedor.sakharov@gmail.com>