chore(vm): Update zkevm test harness (#938)
## What ❔

To make [this](https://github.com/matter-labs/zksync-era/pull/860) PR
more efficient, we need the following PRs to be present:

https://github.com/matter-labs/era-zkevm_test_harness/pull/69
https://github.com/matter-labs/era-zkevm_test_harness/pull/68

This lead to some other dependency changes, so I separate out this
update to ensure that everything is fine

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.