perf: reduce memory consumption of witness generation (#696)
## What ❔

Uses an [updated
zkevm_test_harness](https://github.com/matter-labs/era-zkevm_test_harness/pull/57)
that uses less memory. The updated crate takes callback functions
instead of returning a vector of all results.

We may have to think about how to not keep the database open the whole
time the callbacks are coming in.

## Why ❔

Proving batches currently uses memory proportional to the batch size,
which is not ok for large batches.

---------

Co-authored-by: Fedor Sakharov <fedor.sakharov@gmail.com>
Co-authored-by: AntonD3 <74021421+AntonD3@users.noreply.github.com>
Co-authored-by: Stanislav Breadless <stanislavbezkor@gmail.com>
Co-authored-by: perekopskiy <53865202+perekopskiy@users.noreply.github.com>
Co-authored-by: pompon0 <pompon.pompon@gmail.com>
Co-authored-by: Dustin Brickwood <dustinbrickwood204@gmail.com>
Co-authored-by: Igor Borodin <hatemosphere@protonmail.com>