fix(prover): QoL socket utilization
This PR adds necessary timeout for WVG <-> prover socket communication.
On top of it, also improves the debugging experience by adding logs on
what connections are being established to what IPs.

Without the timeout, we could have WVGs stuck for as long as the
connection takes. In practice, we've observed over 30 minutes.

This will limit such instances to ~1 minute (10 retries * 6 seconds
timeout). Whilst this is a duct-tape, it will do until we get around
implementing the long-term solution.