fix(docs): Set expectations for typo PRs (#1045)
This change is in place to set the expectation for what will happen with
small typo fixes, synonym replacements, etc.

This policy is being added as there are too many PRs opened that either
rename small fragments, rephrase synonyms or fix typos. The goal is to
limit those and focus the energy/ attention on improving the
spellchecker.

This policy eases the "review PR" procedure. It's much easier to say
"no, here's a link why" and close the PR, rather than filtering through
the subjective opinion of the reviewer and figure out if removing some
spaces or replacing "enhance" with "empowers" better fits the
description.

---------

Co-authored-by: Igor Aleksanov <popzxc@yandex.ru>
Co-authored-by: Daniyar Itegulov <ditegulov@gmail.com>