Update main_node.rs
Adding context to errors:

For each method that might return an error, I added context using the context() method:

let node = ZkStackService::new().context("Failed to initialize node")?

This helps produce more informative error messages, making it clear where in the program the problem occurred.
Using ? Here is an example:

In the main() function and in methods such as add_pools_layer(), add_query_eth_client_layer(), etc., we added the operator ?

MainNodeBuilder::new()? // Added `?` operator for error handling
    .add_pools_layer()? // Added `?` for error handling

This allows for concise and clear error handling, propagating errors up the call stack.

It is mentioned that error handling with ? has been added, context to errors has been added, and so on.
This will help a newbie like me understand why these changes were made and how they improve the code.
Overall this tweak makes the code work more understandable in terms of error handling. 
This is my first commit, probably not the most informative, but I also want to take some part in the development of the project I'm in since 2022).