fix: typo and terminology
First change, zk is not indicating it is zksync era or lite, therefore we should make it more clear in the description.

Second change, Consistency in Terminology: Ensure consistency in how you refer to similar items. For example, if you refer to "front end" in one place and "frontend" in another, choose one and stick with it.

Since we use front end multiple times in line 19-21, we would better to change line 10 to front end as well.