feat(api): Implement TxSink abstraction (#1204)
## What ❔

Implements a new abstraction: `TxSink` to be used in `TxSender`.
Right now we have to use either `TxProxy` or master pool (e.g. only one
component), but the code allows invalid configurations. Moreover, mixing
two approaches on the `TxSender` level is unreasonably clunky and
complex.
Having a single abstraction seems to be a better approach.

Additionally, it's a prerequisite for implementing an API layer for the
framework.

⚠️ There are quite a few possibilities for improvements here, but I
suggest not going down the rabbit hole. As long as it's perceived as an
incremental improvement, let's not focus on the adjacent code. We can
always revisit this place next time it becomes problematic.