feat: Remove methods to sign an arbitraty message (#1294)
## What ❔

Removes public APIs that allow signing arbitrary messages.
Replaces https://github.com/matter-labs/zksync-era/pull/482

## Why ❔

The functionality to sign arbitrary messages was used in zkSync Lite
(where it was inherited from), but is not needed for zkSync Era.
Moreover, the implementation is controversial, e.g. we can either add or
not add the Ethereum standard prefix, and without a use case we don't
really have a justification for either option.
Overall, it reduces the surface of public API we commit to support.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.