feat: block revert support for consensus component (#1213)
Reverting blocks in the main node translates to a hard fork of the
consensus blockchain.
This means that every time block reverter is executed, the state of the
consensus has to be reset (aka forked).
To achieve that a `Genesis` of the consensus is introduced. Among other
things, it specifies the number of the fork (incremented at every
forking) and the miniblock number that the fork should start from (set
to the last sealed miniblock at the time of forking). External node
learns the current Genesis from the main node via jsonrpc (eventually
Genesis should be moved to L1 contract, so that hard forks are managed
by governance).

Consensus doesn't support yet synchronization of blocks before the
Genesis first block, therefore for now external node is using jsonrpc
for syncing the blocks before genesis first block.

Fixes BFT-420