fix(eth-sender): adjust the blob tx fees taking into account the current prices (#1399)
## What ❔

It has been noted that simply doubling previous prices in case of blob
transactions may not be a good enough method to get them actually
included. While doubling the prices for replacement transactions is a
_necessary_ thing to do it may not be enough, consider a situation where
an initial blob gas price has been for any reason estimated to be very
low. After a couple of operations of resending the growth of normal fees
would quickly outpace the growth of the blob fees:

```
    "maxFeePerBlobGas": "0x680",
    "maxFeePerGas": "0x6b864500fc0",
    "maxPriorityFeePerGas": "0x2540be4000",
```

But this does not reflect the current effective blob (or normal 1559)
prices onchain. Because at this moment the actual blob prices may be at
`"maxFeePerBlobGas": "0x54c8a469d"` already. So as a solution: take the
`max` of doubled previous prices or the current estimates.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.

(cherry picked from commit fea67fb129601c647908fd758feead4ac8c4255e)