feat(node_framework): Only store each wiring layer once (#1468)
## What ❔

Changes the logic of the framework so that it's OK to call `add_layer`
with the same layer multiple times: the corresponding layer will only be
stored once.

## Why ❔

It may be handy if the same layer (e.g. think gas adjuster) is a
prerequisite for multiple other layers: we can simply bundle them
together without having to worry whether any other layer adds it.

It will be useful for components-based system.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.