feat(node-framework): Add circuit breaker checker layer to framework (#1452)
Key thing to note - no more circuit breaker channel for signalling about
check failure. Instead, make the circuit breaker a part of the other
tasks and return an error instead of a signal throw channel. Motivation
- framework design requires it.

## What ❔
Port circuit breaker checker to the framework

## Why ❔
these changes is a part of the system porting process to the framework

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.