fix(en): Reduce amount of data in snapshot header (#1528)
## What ❔

Reduces the amount of data in `SnapshotHeader` returned by the relevant
RPC method. This missing data is fetched from the main node.

## Why ❔

Right now, the snapshot header returned by JSON-RPC API includes the
entire L1BatchWithMetadata. Most of this information is not used, and
the data itself may be not present in DB! Indeed, new snapshots are
created for the penultimate L1 batch. It is guaranteed to have metadata
calculator run on it (i.e., have a state hash computed), but not the
commitment generator. The corresponding errors have been observed in CI
runs, and in local testing.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.
- [x] Spellcheck has been run via `zk spellcheck`.
- [x] Linkcheck has been run via `zk linkcheck`.