feat: Remove zksync-rs SDK (#1559)
## What ❔

Moves `zksync-rs` SDK to `loadnext` as a module.


`zksync-rs` is deprecated and not really maintained. However, it's a
dependency for `loadnext`, and we don't have capacity to rewrite
`loadnext` to use newer SDK.
Here we simply move what was `zksync-rs` to a module within `loadnext`.

Also, it removes the `sdk` folder completely.

## Why ❔

One less public package to maintain at low cost.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.
- [ ] Linkcheck has been run via `zk linkcheck`.