refactor: TypeId instead of string-based ResourceId (#1518)
`std::any::TypeId` solves the same problem as the current string
identifiers, while using less memory. It eliminates errors that can
occur due to copy-paste or similar naming.

I'm not super happy with how building the name works but it is good
enough in my opinion. I tried to add `const NAME: &'static str` to the
interface but learned that constants can't depend on generic parameters
of the trait impl. At least if the compiler is even a little bit smart
it will concatenate the strings at compile time.